requirements.txt
fastapi
pydantic
opencv-python
pymupdf
python-dotenv
google-generativeai
uvicorn
python-multipart
